+++ styleframe/tests/style_frame_tests.py
@@ -411,6 +411,46 @@ class StyleFrameTest(unittest.TestCase):
                             for row_in_excel, row_in_self in zip(rows_in_excel, rows_in_self)
                             for excel_cell, self_cell in zip(row_in_excel[1:], row_in_self[1:])))
 
+    def test_read_excel_with_style_header_arg_none(self):
+        self.sf = StyleFrame({0: ['A1', 'A2', 'A3', 'A4', 'A5']})
+        self.sf.apply_style_by_indexes(self.sf[self.sf.data_df[0].isin(('A2', 'A5'))], Styler(bold=True))
+        self.export_and_get_default_sheet(save=True)
+        sf_from_excel = StyleFrame.read_excel(TEST_FILENAME, read_style=True, header=None)
+        # making sure content is the same
+        self.assertTrue(all(list(self.sf[col]) == list(sf_from_excel[col])[1:] for col in self.sf.columns))
+
+        rows_in_excel = list(sf_from_excel.data_df.itertuples())
+        rows_in_self = self.sf.data_df.itertuples()
+
+        # making sure styles are the same
+        self.assertTrue(all(excel_cell.style == self_cell.style
+                            for row_in_excel, row_in_self in zip(rows_in_excel[1:], rows_in_self)
+                            for excel_cell, self_cell in zip(row_in_excel[1:], row_in_self[1:])))
+
+    def test_read_excel_rows_height(self):
+        self.sf.set_row_height(rows=1, height=25)
+        self.sf.set_row_height(rows=2, height=15)
+        self.export_and_get_default_sheet(save=True)
+        sf_from_excel = StyleFrame.read_excel(TEST_FILENAME, read_style=True)
+
+        # Assert the number of rows with height is the length of our data plus 1 for headers row
+        self.assertEqual(len(sf_from_excel._rows_height), len(self.sf) + 1)
+        self.assertEqual(sf_from_excel._rows_height[1], 25)
+        self.assertEqual(sf_from_excel._rows_height[2], 15)
+        self.assertEqual(sf_from_excel._rows_height[3], None)
+        self.assertEqual(sf_from_excel._rows_height[4], None)
+
+    def test_read_excel_columns_width(self):
+        self.sf.set_column_width(columns='a', width=25)
+        self.sf.set_column_width(columns='b', width=15)
+        self.export_and_get_default_sheet(save=True)
+        sf_from_excel = StyleFrame.read_excel(TEST_FILENAME, read_style=True)
+
+        # Assert the number of rows with height is the length of our data plus 1 for headers row
+        self.assertEqual(len(sf_from_excel._columns_width), len(self.sf.columns))
+        self.assertEqual(sf_from_excel._columns_width['a'], 25)
+        self.assertEqual(sf_from_excel._columns_width['b'], 15)
+
     def test_read_excel_template_equal_boundaries(self):
         template_sf = StyleFrame(
                         obj={